Using GroupModel for Group and old way Ticket for Tickets, Test further.
Testing PDO::FETCH_CLASS to hydrate the class. At this point i am using two files: Group and GroupModel as the Group CLASS (One for the getters and setters, the other for methods and hydration) but i still have to reproduce the same parameters, wish, i think is totally unnecessary. TO REVIEW